refactor: ajustes migracoes e rotas

+++ db/migrate/20250628145735_create_a_cidades.rb
@@ -2,6 +2,7 @@
 
 class CreateACidades < ActiveRecord::Migration[7.2]
   def up
+   unless table_exists?(:a_cidades)
     create_table :a_cidades do |t|
       t.string :descricao
       t.references :a_estado
@@ -11,9 +12,10 @@ class CreateACidades < ActiveRecord::Migration[7.2]
       t.datetime :deleted_at
       t.timestamps
     end
+   end
   end
 
   def down
-    drop_table :a_cidades
+    drop_table :a_cidades if table_exists?(:a_cidades)
   end
 end
