+++ db/migrate/20250628150157_create_a_sexos.rb
@@ -2,6 +2,7 @@
 
 class CreateASexos < ActiveRecord::Migration[7.2]
   def up
+   unless table_exists?(:a_sexos)
     create_table :a_sexos do |t|
       t.string :descricao
       t.string :sigla
@@ -11,9 +12,10 @@ class CreateASexos < ActiveRecord::Migration[7.2]
       t.datetime :deleted_at
       t.timestamps
     end
+   end
   end
 
   def down
-    drop_table :a_sexos
+    drop_table :a_sexos if table_exists?(:a_sexos)
   end
 end
